Minor clean up
Remove old BattleManager and replaced with BattleManagerNew.
Remove Singleton of DicePoolManager and added reference to BattleManager.
Removed unused/outdated Dice scripts
