--- Assets/BattleEntity.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8d8ef29576d4d4e4da8bbc2ecf89d842
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
