--- Assets/BattleManagerNew.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 432cd95e2fae27b40a41ae0852904e12
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
