--- Assets/Scripts/DiceMoveManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 5d7a9536c5b9849ba8b5fc0da66e80b6
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
