--- Assets/Scripts/DiceRollAnimation.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2942d96e6234e4212a7e55e418e58050
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
