--- Assets/Scripts/DiceRollTest.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9742719e81e3545ad848a913fe516f1d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
